@@ -24,7 +24,6 @@
     "ts-jest": "^29.1.0",
     "ts-loader": "^7.0.5",
     "ts-node": "^8.10.2",
-    "typescript": "^3.9.5"
-  },
-  "dependencies": {}
+    "typescript": "^5.0.4"
+  }
 }
